requirements.txt
streamlit>=1.32.0
duckdb>=0.10.0
pandas>=2.0.0
numpy>=1.26.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
pyarrow>=14.0.0
requests>=2.31.0

# Install torch separately with CUDA support:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
